Handle frame type and background with CSS.
As Gtk3 ruined the GtkFrame shadowing and deprecated the background
property, use a dynamic CSS rule to handle them. The frame is now
represented by a generic container (GtkGrid) and a CSS rule is applied
to it. Each time the configuration changes, the rule is reconstructed
and reapplied.

A new plain (i.e.: solid) frame type is now featured.

Border shadowing uses semi-transparent grayscale colors to respect the
original background color.

It is now possible to configure whether the frame should occupy the
whole applet area or the minimum space to afford its contents.

Panel orientation is monitored to apply padding on the proper axis.